requirements.txt
# SVD图像压缩工具依赖

# 核心依赖
numpy>=1.21.0
Pillow>=8.0.0

# 图像处理和评估
opencv-python>=4.5.0  # 用于图像处理和质量评估
scikit-image>=0.19.0  # 用于SSIM等质量指标

# 可视化（可选）
matplotlib>=3.5.0     # 用于对比可视化

# 开发工具
tqdm>=4.62.0          # 进度条（用于批量处理）
argparse              # 命令行参数解析（Python标准库）